fix documentation dialog

+++ src/main/java/inspections/DeletePatternInstanceQuickFix.java
@@ -6,19 +6,16 @@ import com.intellij.openapi.project.Project;
 import org.jetbrains.annotations.Nls;
 import org.jetbrains.annotations.NotNull;
 import storage.PluginState;
-import storage.ProjectDetails;
 import storage.ProjectPersistedState;
+import storage.ProjectsPersistedState;
 
 public class DeletePatternInstanceQuickFix implements LocalQuickFix {
 
     public static final String QUICK_FIX_NAME = "Remove pattern instance";
     private String patternInstanceId;
-    private ProjectPersistedState projectPersistedState;
 
     public DeletePatternInstanceQuickFix(String patternInstanceId){
         this.patternInstanceId = patternInstanceId;
-        ProjectDetails projectDetails = PluginState.getInstance().getProjectDetails();
-        projectPersistedState = projectDetails.getActiveProjectPersistedState();
     }
 
     @Nls(capitalization = Nls.Capitalization.Sentence)
@@ -30,6 +27,11 @@ public class DeletePatternInstanceQuickFix implements LocalQuickFix {
 
     @Override
     public void applyFix(@NotNull Project project, @NotNull ProblemDescriptor descriptor) {
-        projectPersistedState.deletePatternInstance(patternInstanceId);
+        ProjectsPersistedState projectsPersistedState = PluginState.getInstance().getState();
+
+        if (projectsPersistedState != null) {
+            ProjectPersistedState projectPersistedState = projectsPersistedState.getProjectState(project.getName());
+            projectPersistedState.deletePatternInstance(patternInstanceId);
+        }
     }
 }
